% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_languages.R
\name{available_languages}
\alias{available_languages}
\title{List supported languages}
\usage{
available_languages()
}
\value{
Returns the list of supported languages.
}
\description{
\code{available_languages} list all supported languages.
}
