% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_transformer.R
\name{init_transformers}
\alias{init_transformers}
\title{Initialize transformers}
\usage{
init_transformers(
  condaenv,
  model_en_zh = "Helsinki-NLP/opus-mt-en-zh",
  model_zh_en = "Helsinki-NLP/opus-mt-zh-en"
)
}
\arguments{
\item{condaenv}{path of the conda environment to use.}

\item{model_en_zh}{model for English to Chinese translation.}

\item{model_zh_en}{model for Chinese to English translation.}
}
\value{
the transformer object.
}
\description{
\code{init_transformers} initialize transformers for translation.
}
