% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hash_from_dictionary.R
\name{make_hash_from_dictionary}
\alias{make_hash_from_dictionary}
\title{Make hash table from dictionary.}
\usage{
make_hash_from_dictionary(dictionary, from_column, to_column)
}
\arguments{
\item{dictionary}{a data frame contains a customerized dictionary.}

\item{from_column}{the column name of the source terms.}

\item{to_column}{the column name of the target terms.}
}
\value{
a hash table of the dictionary.
}
\description{
\code{make_hash_from_dictionary} makes the hash table from a customerized dictionary.
}
