% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_translate_df_en_zh.R
\name{med_translate_df_en_zh}
\alias{med_translate_df_en_zh}
\title{Translate a data frame form English to Chinese.}
\usage{
med_translate_df_en_zh(df, transformer, col_names = NULL, dict_map = NULL)
}
\arguments{
\item{df}{a data frame to translate.}

\item{transformer}{functions to transform the string, initialized by \code{init_transformers}.}

\item{col_names}{a vector of column names to translate. If \code{NULL}, all columns will be translated.}

\item{dict_map}{a named vector of customerized dictionary.}
}
\value{
a new data frame with Chinese translation.
}
\description{
\code{med_translate_df_en_zh} translates a data frame from English to Chinese.
}
