% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_translate_df_zh_en.R
\name{med_translate_df_zh_en}
\alias{med_translate_df_zh_en}
\title{Translate a data frame form Chinese to English.}
\usage{
med_translate_df_zh_en(df, transformer, col_names = NULL, dict_map = NULL)
}
\arguments{
\item{df}{a data frame to translate.}

\item{transformer}{the transformer object, initialized by \code{init_transformer}.}

\item{col_names}{a vector of column names to translate. If \code{NULL}, all columns will be translated.}

\item{dict_map}{a named vector of customerized dictionary.}
}
\value{
a new data frame with English translation.
}
\description{
\code{med_translate_df_zh_en} translates a data frame from Chinese to English.
}
