% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_translate_col_en_zh.R
\name{med_translate_col_en_zh}
\alias{med_translate_col_en_zh}
\title{Translate a data column form English to Chinese.}
\usage{
med_translate_col_en_zh(col, transformer, dict_hash = NULL)
}
\arguments{
\item{col}{a data column to translate.}

\item{transformer}{the transformer object, initialized by \code{init_transformer}.}

\item{dict_hash}{a customerized dictionary hash table, made by \code{make_hash_from_dictionary}.}
}
\value{
a new data frame with Chinese translation, and confidences.
}
\description{
\code{med_translate_col_en_zh} translates a data column from English to Chinese.
}
