% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_translate_col_zh_en.R
\name{med_translate_col_zh_en}
\alias{med_translate_col_zh_en}
\title{Translate a data column form Chinese to English.}
\usage{
med_translate_col_zh_en(col, transformer, dict_hash = NULL)
}
\arguments{
\item{col}{a data column to translate.}

\item{transformer}{functions to transform the string, initialized by \code{init_transformers}.}

\item{dict_hash}{a customerized dictionary hash table, made by \code{make_hash_from_dictionary}.}
}
\value{
a new data frame with English translation, and confidence.
}
\description{
\code{med_translate_col_zh_en} translates a data column from Chinese to English.
}
