% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_translate_str_en_zh.R
\name{med_translate_str_en_zh}
\alias{med_translate_str_en_zh}
\title{Translate a string form English to Chinese.}
\usage{
med_translate_str_en_zh(str, transformer, dict_hash = NULL)
}
\arguments{
\item{str}{a string to be translated.}

\item{transformer}{functions to transform the string, initialized by \code{init_transformers}.}

\item{dict_hash}{a customerized dictionary hash table, made by \code{make_hash_from_dictionary}.}
}
\value{
Chinese translation with confidence.
}
\description{
\code{med_translate_str_en_zh} translates a string from English to Chinese.
}
